# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API server. (This is a practice of fumiyasac.)",
        "title": "Sample API",
        "contact": {
            "name": "fumiyasac",
            "url": "https://qiita.com/fumiyasac@github",
            "email": "just1factory@gmail.com"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/top": {
            "get": {
                "description": "TOP用の挨拶メッセージとこのAPIに関する概要や現行バージョン・更新履歴を表示する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "このAPIに関するメッセージを表示する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.TopMessageFactory"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "userテーブルに登録されているデータを全件取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザー一覧を表示する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.UserItemsFactory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APIErrorMessageFactory"
                        }
                    }
                }
            },
            "post": {
                "description": "userテーブルへ新規データを1件登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザーデータを1件新規登録する",
                "parameters": [
                    {
                        "description": "ユーザー名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "パスワード",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "メールアドレス",
                        "name": "mailaddress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APISuccessMessageFactory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APIErrorMessageFactory"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "userテーブルに登録されているデータのうちパラメーター(id)に該当するものを1件取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザーデータを1件表示する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.SingleUserFactory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APIErrorMessageFactory"
                        }
                    }
                }
            },
            "put": {
                "description": "userテーブルに登録されているデータのうちパラメーター(id)に該当するものを1件更新する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザーデータを1件更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ユーザー名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "パスワード",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "メールアドレス",
                        "name": "mailaddress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APISuccessMessageFactory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APIErrorMessageFactory"
                        }
                    }
                }
            },
            "delete": {
                "description": "userテーブルに登録されているデータのうちパラメーター(id)に該当するものを1件削除する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザーデータを1件削除する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APISuccessMessageFactory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/factories.APIErrorMessageFactory"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "factories.APIErrorMessageFactory": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "factories.APISuccessMessageFactory": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "string"
                }
            }
        },
        "factories.SingleUserFactory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mailaddress": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "factories.TopMessageFactory": {
            "type": "object",
            "properties": {
                "announcements": {
                    "type": "array",
                    "items": {
                        "type": "\u0026{entities TopAnnouncement}"
                    }
                },
                "detail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "factories.UserItemsFactory": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/factories.SingleUserFactory"
                    }
                }
            }
        }
    }
}